LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/7.9.2_Examples/S7.9.2_A1_T6-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Unexpected keyword 'else'
S7.9.2_A1_T6

PASS Expected parsing failure

TEST COMPLETE
